fix: Selected color is incorrectly set to black when selecting text
Unfortunately it seems like there's a bug in SwiftUI that causes the currently selected color to be set to black if the user selects some text in the app when the color picker is open. This change works around that by closing the color picker in scenarios when it might occur and disables text selection where not aboslutely necessary. Hopefully SwiftUI will fix this in the future and we can remove the workaround (see https://github.com/inseven/symbolic/issues/177).